Refactoring structure to keep order
