+++ src/main/kotlin/Command/base/Cart.kt
@@ -6,18 +6,15 @@ import Decorator.littlekai.ingredients.Chicken
 import Decorator.littlekai.ingredients.Peanuts
 import Decorator.littlekai.ingredients.Pork
 import Decorator.littlekai.ingredients.Tuna
-import Decorator.littlekai.noodles.EggNoodles
-import Decorator.littlekai.noodles.UdonNoodles
-import Decorator.littlekai.noodles.WheatNoodles
 import Decorator.littlekai.sauces.BittersweetSauce
 import Decorator.littlekai.sauces.RedPepperSauce
 import Decorator.littlekai.sauces.SateSauce
 import Decorator.littlekai.sauces.TeriyakiSauce
 import Factory.base.NoodlesFactory
-import strategy.GeneralStrategy
-import strategy.IvaStrategy
-import strategy.ReducedStrategy
-import strategy.SuperReducedStrategy
+import Strategy.base.IvaStrategy
+import Strategy.strategies.GeneralStrategy
+import Strategy.strategies.ReducedStrategy
+import Strategy.strategies.SuperReducedStrategy
 import java.util.*
 
 open class Cart(private val scanner: Scanner,
