+++ src/main/kotlin/Strategy/base/IvaStrategy.kt
@@ -1,0 +1,5 @@
+package Strategy.base
+
+interface IvaStrategy {
+  fun applyIva(cost: Double) : Double
+}
