+++ src/main/kotlin/Strategy/strategies/GeneralStrategy.kt
@@ -1,4 +1,6 @@
-package strategy
+package Strategy.strategies
+
+import Strategy.base.IvaStrategy
 
 class GeneralStrategy : IvaStrategy {
